Backend/requirements.txt
djangorestframework
django-cors-headers
Pillow
mysqlclient
reportlab